Fix: Use `OSError.from_u32!` instead of non-existent `from_u64!`.
The `from_u64!` function worked when `OSError` was an `:enum`, but it
stopped being an enum when we needed to support using different
values for certain errors on different platforms.

This code has been broken since that change away from being an enum.

This commit fixes the code by using the newly introduced
`from_u32!` function that is now provided by `OSError`.